broker/requirements.txt
# file: broker/requirements.txt
fastapi
uvicorn[standard]
pydantic
pyserial
esptool